requirements.txt
tensorflow>=2.12.0
numpy>=1.23.0
Pillow>=9.5.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
pyyaml>=6.0
streamlit>=1.28.0
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
